only active courses displayed on student page, courses with unspecified visibility treated like active
